Release v1.4.0 (#15)
Non-breaking changes:
- New APIs:
    - SDK now supports push notifications.
        - Pass the push notification device token to register the device with `set(notificationDeviceToken: String)`
        - Check that the tapped notification is an Octopus notification with `static func isAnOctopusNotification(notification: UNNotification) -> Bool`
        - Tell the UI to handle the tapped notification with the `notificationResponse` binding passed to `OctopusHomeScreen`
    - Track some custom events with `track(customEvent: CustomEvent) async throws`
    - Changed the look of the Octopus home screen with the params `navBarLeadingItem` and `navBarPrimaryColor`.

- Internal changes:
    - Button to see the parent post when looking at a comment after clicking on an internal (or push) notification
    - Users can start to write their comments or replies before being invited to create their profile
    - Limiting the min length of the text of a post if there is no poll associated
    - UI/UX improvments

+++ Sources/OctopusCore/Topics/TopicsRepository.swift
@@ -25,12 +25,14 @@ public class TopicsRepository: InjectableObject, @unchecked Sendable {
     @Published public private(set) var topics: [Topic] = []
 
     private let topicsDatabase: TopicsDatabase
+    private let networkMonitor: NetworkMonitor
     private let remoteClient: OctopusRemoteClient
     private var storage: Set<AnyCancellable> = []
 
     init(injector: Injector) {
         topicsDatabase = injector.getInjected(identifiedBy: Injected.topicsDatabase)
         remoteClient = injector.getInjected(identifiedBy: Injected.remoteClient)
+        networkMonitor = injector.getInjected(identifiedBy: Injected.networkMonitor)
 
         topicsDatabase.topicsPublisher()
             .replaceError(with: [])
@@ -41,6 +43,7 @@ public class TopicsRepository: InjectableObject, @unchecked Sendable {
     }
 
     public func fetchTopics() async throws {
+        guard networkMonitor.connectionAvailable else { throw ServerCallError.noNetwork }
         let response = try await remoteClient.octoService.getTopics(authenticationMethod: .notAuthenticated)
         let topics = response.topics.compactMap { Topic(from: $0) }
         try await topicsDatabase.deleteAll()
